Now everything is generalized with functions. The whole thing is working partially. And, also, we need to approach the problem from a different perspective: Whenever we send a message, then we can also include the parameter of the id of the last message posted in the chat. In this way, we are going to retrieve every message sent since the last time we retrieved them, including the message we are sending now.
Also, we will  include the extra parameter of the id in the retrieveMessages function.
We need this function to be executed with a fixed rate, and each time it is executed it will put itself in the task queue as well with a timeout, having in account that maybe we will use the same handler callback for both cases.

Important thing: After every chat operation we will check the state if necessary and retrieve the new messages:
- erase all messages: we will erase all messages in the database and then check for new messages in case other users could have sent them. Also, we will perform asynchronous tasks, that is, we will have to check if there is any message before the last id, in order to know if the database messages have been erased. But, of course, this would be necessary just in case we want to have a control to delete all messages in the database. It would be much better a control to delete just old messages. Also, we could have a function/control to clear the former messages in the chat for this user. Then, we will erase then and we have to include a field like <data-id> in each message, and after ordering the clearing of messages, we will maintain a variable with the value of the last erased message from the chat or the first one after that one.

So, we will maintain the id of the last id retrieved, and the id of the first id from which we want to load from. Just making a db query with these values should be fine. Also, loading messages will be included in the other chat queries like post new message, etc, and in that case we will put a timeout to launch the next task. If a task is running, we will never run  retrieveMessages, instead we will check it and discard the function.

=======================================================
Every task we run will be waiting until the last task has finished, and then will run. It would never occur that one task overlaps other one, but never know, this is asynchronous.  
Every time we put a sendMessage task in the table, we will first empty the table. Afterwards we will put a new 
Every time we put a login, logout or register task in the table, we will first empty the table.
Every time we put a retrieveMessages task in the table, this will be automatically done at the end of every task handler, so we don't really need to bother about this.
=========================================

About the functionality of deleting every messages, I am not very sure, but maybe I would like to implement other one just for the admin user in order to delete old messages.