Yayy!! We could say that the chat functionality itself is working pretty well now, for guest users. We commit at this point since now, we will try to make first the login and register work properly, and after that, make sure that the chat works also for registered users when logged in.
After that, it would be nice to implement some 'forgotten password' with ajax. And after that... I guess if we feel like that maybe we could try to put smileys... buf we will see haha. We are thinking too long afar now! So, as I said just now, we go now for login and register function.