Also, we should use the password always coded with sha1 or md5. This subject is delaying, would be good to handle it soon. At this point the login, register and chat work fine. One nice thing to do would be add a feature like guest user can choose between random user name or write one name him/herself. Always checking that the name is not in the database. Good thing is: if the name does not meet the requeriments for an user name, there we have a shortcut for those cases.
After that, it would be nice to implement some 'forgotten password' with ajax. 

And it would be nice to implement the functionality for the text style. Shouldn't be that difficult hehe :P
And after that... I guess if we feel like that maybe we could try to put smileys... buf we will see haha. We are thinking too long afar now! So, as I said just now, we go now for login and register function.