We implemented already a system with which a guest user can change the user name.
Now, we are going to implement some 'forgotten password' with ajax. This link  should  appear in the chat window as well as in the login window. And it will load a new page forgotpassword or sth, where the email with a link will be sent to the user, checking first that the email is in the database.

And it would be nice to implement the functionality for the text style. Shouldn't be that difficult hehe :P
And after that... I guess if we feel like that maybe we could try to put smileys... buf we will see haha. We are thinking too long afar now! So, as I said just now, we go now for login and register function.