We implemented already a system with which a guest user can change the user name in the chat. We implemented the forgotten password as well. So, now we commit here since we are going to attempt the confirmation email when registering. We are not going to add the functionality of managing the account, nor the profile image, yet. We could do that in a next version.
The thing to think about could be whether the php objects we create are automatically destroyed.

We will attach this next thing:
And it would be nice to implement the functionality for the text style. Shouldn't be that difficult hehe :P
And after that... I guess if we feel like that maybe we could try to put smileys... buf we will see haha. We are thinking too long afar now! So, as I said just now, we go now for login and register function.

Now, if we were to think about changing this chat to json, we will have to do things with node.js I guess, and mongodb. Not a bad idea I guess.